# docs/swagger.yaml
definitions:
...
  api.DrbdEnableSetting:
    properties:
      enabledrbd:
        type: boolean
      state:
        description: RemoveReservedRsp
        type: string
      version:
        type: string
    type: object
  api.DrbdEnableSettingReqBody:
    properties:
      enable:
        type: boolean
    type: object
  api.DrbdEnableSettingRspBody:
    properties:
      data:
        $ref: '#/definitions/api.DrbdEnableSetting'
    type: object
...
  api.MigrateVolumeGroupVolumeInfos:
    properties:
      volumeGroupVolumeInfos:
        description: VolumeGroupVolumeInfo
        items:
          $ref: '#/definitions/api.VolumeGroupVolumeInfo'
        type: array
      volumeName:
        description: Name
        type: string
    type: object
...
  api.Pagination:
    properties:
      page:
        description: 当前页索引，从 1 开始，为 0 时，会自动重置为默认值 constants.DefaultPage
        type: integer
      pageSize:
        description: |-
          每页数据量，为 -1 时表示查询全部，为 0 时会重置为默认值
          constants.DefaultPageSize
        type: integer
      pages:
        description: 总页数
        type: integer
      total:
        description: 总共有多少条目，请求时可以不用传递
        type: integer
    type: object
  api.RspFailBody:
    properties:
      description:
        type: string
      errcode:
        type: integer
    type: object
...
  api.Volume:
    properties:
      convertible:
        description: Convertible 转换高可用模式
        type: boolean
      createTime:
        description: createTime 创建时间
        type: string
      name:
        description: local volume name 名称
        type: string
      pvcName:
        description: PersistentVolumeClaimName is the name of the associated PVC 绑定PVC
        type: string
      pvcNamespace:
        description: PersistentVolumeClaimNamespace is the namespace of the associated
          PVC 命名空间
        type: string
      replicaNumber:
        description: replica number 副本数
        type: integer
      requiredCapacityBytes:
        description: size 容量
        type: integer
      state:
        description: local volume state 状态
        type: string
      volumeGroup:
        description: VolumeGroup is the group name of the local volumes. It is designed
          for the scheduling and allocating. 磁盘组
        type: string
    type: object
  api.VolumeConvertInfo:
    properties:
      replicaNum:
        type: integer
      volumeName:
        type: string
    type: object
  api.VolumeConvertReqBody:
    properties:
      volumeName:
        type: string
    type: object
  api.VolumeConvertRspBody:
    properties:
      data:
        $ref: '#/definitions/api.VolumeConvertInfo'
    type: object
...
  api.VolumeGroupVolumeInfo:
    properties:
      nodeNames:
        description: NodeNames
        items:
          type: string
        type: array
      state:
        description: State
        type: string
      volumeName:
        description: VolumeName
        type: string
    type: object
  api.VolumeList:
    properties:
      items:
        description: |-
          // volumes
          VolumeItemsList VolumeItemsList `json:"items,omitempty"`
        items:
          $ref: '#/definitions/api.Volume'
        type: array
      pagination:
        $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
  api.VolumeMigrateInfo:
    properties:
      selectedNode:
        type: string
      srcNode:
        type: string
      volumeName:
        type: string
    type: object
  api.VolumeMigrateOperation:
    properties:
      endTime:
        description: EndTime 迁移结束时间
        type: string
      name:
        description: VolumeMigrateName 迁移CRD名称
        type: string
      sourceNode:
        description: SourceNode 迁移源节点
        type: string
      startTime:
        description: StartTime 迁移开始时间
        type: string
      state:
        description: State 迁移状态
        type: string
      targetNode:
        description: TargetNode 迁移目的节点
        type: string
      volumeName:
        description: VolumeName 迁移卷名称
        type: string
    type: object
  api.VolumeMigrateReqBody:
    properties:
      selectedNode:
        type: string
      srcNode:
        type: string
    type: object
  api.VolumeMigrateRspBody:
    properties:
      data:
        $ref: '#/definitions/api.VolumeMigrateInfo'
    type: object
  api.VolumeOperationByVolume:
    properties:
      items:
        description: |-
          // VolumeMigrateOperationItemsList
          VolumeMigrateOperationItemsList VolumeMigrateOperationItemsList `json:"items,omitempty"`
          VolumeMigrateOperations
        items:
          $ref: '#/definitions/api.VolumeMigrateOperation'
        type: array
      volumeName:
        description: VolumeName
        type: string
    type: object
...
  api.VolumeReplica:
    properties:
      devicePath:
        description: |-
          DevicePath is a link path of the StoragePath of the volume replica,
          e.g. /dev/LocalStorage_PoolHDD/pvc-fbf3ffc3-66db-4dae-9032-bda3c61b8f85
        type: string
      name:
        description: replica name
        type: string
      nodeName:
        description: NodeName is the assigned node where the volume replica is located
          节点
        type: string
      requiredCapacityBytes:
        description: RequiredCapacityBytes 容量
        type: integer
      state:
        description: replica state
        type: string
      storagePath:
        description: StoragePath is a real path of the volume replica, like /dev/sdg.
        type: string
      synced:
        description: Synced is the sync state of the volume replica, which is important
          in HA volume 同步状态
        type: boolean
    type: object
  api.VolumeReplicaList:
    properties:
      items:
        description: VolumeReplicas
        items:
          $ref: '#/definitions/api.VolumeReplica'
        type: array
      volumeName:
        description: volume name
        type: string
    type: object
  api.YamlData:
    properties:
      data:
        description: yaml data
        type: string
    type: object
info:
  contact: {}
paths:
...
  /settings/highavailabilitysetting:
    get:
      consumes:
      - application/json
      description: get DRBDSettingGet
      parameters:
      - description: enabledrbd
        in: path
        name: enabledrbd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DrbdEnableSetting'
      summary: 摘要 获取高可用设置
      tags:
      - Setting
  /settings/highavailabilitysetting/drbd:
    post:
      consumes:
      - application/json
      description: post EnableDRBDSetting
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DrbdEnableSettingReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DrbdEnableSettingRspBody'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 高可用设置
      tags:
      - Setting
...
  /volumes/volume/{name}/yaml:
    get:
      consumes:
      - application/json
      description: get VolumeYamlGet
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.YamlData'
      summary: 摘要 获取数据卷yaml信息
      tags:
      - Volume
  /volumes/volumemigrateoperation/{volumeName}/volumelist:
    get:
      consumes:
      - application/json
      description: get VolumeListWithSameVolumeGroup
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MigrateVolumeGroupVolumeInfos'
      summary: 摘要 获取同一数据卷组中包含数据卷名称列表信息
      tags:
      - Volume
...
  /volumes/volumeoperation/{volumeName}/convert:
    post:
      consumes:
      - application/json
      description: post VolumeConvertOperation
      parameters:
      - description: volumeName
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.VolumeConvertReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.VolumeConvertRspBody'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 指定数据卷转换操作
      tags:
      - Volume
  /volumes/volumeoperation/{volumeName}/migrate:
    post:
      consumes:
      - application/json
      description: post VolumeMigrateOperation body i.g. body { SrcNode string ,SelectedNode
        string}
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: reqBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.VolumeMigrateReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.VolumeMigrateRspBody'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 指定数据卷迁移操作
      tags:
      - Volume
  /volumes/volumeoperation/{volumeOperationName}/yaml:
    get:
      consumes:
      - application/json
      description: get VolumeOperationYamlGet
      parameters:
      - description: volumeOperationName
        in: path
        name: volumeOperationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.YamlData'
      summary: 摘要 获取数据卷操作记录yaml信息
      tags:
      - Volume
  /volumes/volumeoperations/{volumeName}:
    get:
      consumes:
      - application/json
      description: get VolumeOperation
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: volumeMigrateName
        in: query
        name: volumeMigrateName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.VolumeOperationByVolume'
      summary: 摘要 获取指定数据卷操作记录信息（目前仅包含迁移运维操作）
      tags:
      - Volume
  /volumes/volumereplica/{volumeReplicaName}/yaml:
    get:
      consumes:
      - application/json
      description: get VolumeReplicaYaml
      parameters:
      - description: volumeReplicaName
        in: path
        name: volumeReplicaName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.YamlData'
      summary: 摘要 查看指定数据卷副本yaml信息
      tags:
      - Volume
  /volumes/volumereplicas/{volumeName}:
    get:
      consumes:
      - application/json
      description: list volumes
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: volumeReplicaName
        in: query
        name: volumeReplicaName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.VolumeReplicaList'
      summary: 摘要 获取指定数据卷的副本列表信息
      tags:
      - Volume
  /volumes/volumes:
    get:
      consumes:
      - application/json
      description: list Volume
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: volumeName
        in: query
        name: volumeName
        type: string
      - description: state
        in: query
        name: state
        type: string
      - description: namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.VolumeList'
      summary: 摘要 获取数据卷列表信息
      tags:
      - Volume
  /volumes/volumes/{name}:
    get:
      consumes:
      - application/json
      description: get Volume
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Volume'
      summary: 摘要 获取指定数据卷基本信息
      tags:
      - Volume
swagger: "2.0"